+++ src/langrila/claude/__init__.py
@@ -1,5 +1,5 @@
+from .assembly import Claude
 from .llm.chat import AnthropicChatModule
 from .llm.function_calling import AnthropicFunctionCallingModule
 from .message import ClaudeMessage
-from .model import Claude
 from .tools import ToolConfig, ToolParameter, ToolProperty
