+++ src/langrila/gemini/__init__.py
@@ -1,4 +1,4 @@
+from .assembly import Gemini
 from .llm.chat import GeminiChatModule
 from .llm.function_calling import GeminiFunctionCallingModule
-from .model import Gemini
 # from .message.gemini_api import GeminiMessage
