+++ src/langrila/gemini/assembly.py
