+++ src/langrila/openai/assembly.py
